requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Login>=0.6
Flask-Mail>=0.10
Flask-WTF>=1.2
Flask-Bcrypt>=1.0
Flask-Babel>=4.0
itsdangerous>=2.1
PyMySQL>=1.1
python-dotenv>=1.0
email-validator>=2.0
